shared/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orw-common"
version = "0.1.0"
description = "OpenRadiusWeb Common Library"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "nats-py>=2.6",
    "structlog>=23.2",
]

[tool.setuptools.packages.find]
include = ["orw_common*"]
